📝 docs(social-recovery.md): add explanation and image for initiating funds spending via social recovery in multisig setup
The commit adds an explanation of how to initiate the spending of funds via social recovery in a multisig setup. It mentions that this can be done either by a designated signer or by following the social recovery path. It also states that this is only possible after the specified timelock duration has elapsed and requires the authorization of at least one of the two designated signers for recovery.

An image is also added to illustrate the process.

![image](https://github.com/smartvaults/smartvaults/assets/88842056/82d15056-8679-4ba0-8bfe-2095ffccb6c9)